+++ lib/fetch_api.dart
@@ -1,3 +1,9 @@
+/// The Fetch API provides an interface for fetching resources
+/// (including across the network). It will seem familiar to anyone who has used
+/// `XMLHttpRequest`, but the new API provides a more powerful and flexible
+/// feature set.
+library fetch_api;
+
 export 'src/abort_controller.dart';
 export 'src/abort_signal.dart';
 export 'src/fetch.dart';